pyproject.toml
[project]
name = "google-adk-extras"
version = "0.1.0"
description = "Custom implementations of Google ADK services (Session, Artifact, Memory) with multiple storage backends"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pymongo>=4.0.0",
    "redis>=4.0.0",
    "pyyaml>=6.0.0",
    "typing-extensions>=4.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "google-adk>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.15.0",
    "google-genai>=1.0.0",
]
s3 = [
    "boto3>=1.0.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
